added Functions to add together odds that customer will buy lemonade

+++ LemonadeStand/Day.cs
@@ -7,7 +7,7 @@ namespace LemonadeStand
 {
     public class Day
     {
-        int Time;
+       float Time;
         public Weather Weather;
